+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 - Fix an issue with some BetterTTV emotes.
 - Improve error handling for unavailable channels.
 
+### ⚙️ Internal
+
+- Remove service worker cache.
+
 # 0.1.0
 
 ### 🎉 Initial release